remove difficult scenarios from dPendulum evolver
it is quite difficult to solve the double pendulum so that it not only
balances inverted pendulums, but so that it also swings up hanging
pendulums then balances them.

i've removed this set of test scenarios and scenarios in which the
pendulums are pushed very hard at the beginning of the scenario.

this change increases the speed with which solutions to the limited
domain are found (due to simulation time) and improves results since
they are not a compromise between a good solution to the solvable parts
and a bad solution to the difficult ones.