create callback for pushing poles of d. pendulum
instead of pushing the poles in the double pendulum demo, a callback is
now used for pushing the poles multiple times during the experiment.
this also allows more testing to occur in one test round, removing the
need to test each network several times, thus speeding up the testing
process.

this will also make it easier to program user interaction with the
graphical demos.